Add most unit models, some battle controller
This adds all of the unit models for infantry, artillery,
vehicles, and aircraft, as well as some basic battle controller
work done not taking into account much more beyond basic
hits. Should all work, with main as a basic testing program.

Should be able to use this as the backend for a web
based tool that will use these models when units picked
in a selector.