Update battle controller, add tests
This adds a bunch of different things:
* Updates to some units which were incorrect
* Change battle controller to use new Casualties object
* Added tests for target selection and a few select units
* Added toml file for the project
* Added facilities and their impact on battles

Some stuff to clean up in the future is the tests right now
need to be run with python3 -m pytest from the src directory.
That should be able to be run with pytest from the top level,
as that would be more proper. Will do this in a refactor for
later, but for now this will suffice for a simple way to
do things.