Basic web stuff and some bug fixes
This is mostly just to get a basic flask frontend
working. All it does it allow the user to pick a
bunch of units from a small subset of available
units, mostly as a proof of concept so I can do more
here later.

A couple of other small bug fixes that came to light,
in the battle controller and in the terrain model.

Lastly, added a requirements file with the specific
versions I am using for testing so that the same environment
can be deployed later on easily.