Add ArmyResolver tests
A bunch of tests that are strictly working on the
default hit functionality, making sure units are being
taken in an appropriate order. Also making sure that
units with extra health are taken first. A couple of
small changes to the weight formula to give units that
first strike or blitz an additional point for determining
casulaties.