Add target selection, fix pip math
This handles a very basic type of target selection
and should be optimal as long as someone doesn't
get too many and get them accidentally downgraded
when there was a better combo that wasn't the first.

Also fixed a major issue with the pip math. tmin should have
been being added, not subtracted, since it was always going
to be a number 0 or less, so subtracting would be add more
to the adjustment, and not subtracting like it should.
Added a small test that verifies this for the infrantry
unit in the mountain, will add more later.