Add factory creator for units, add all units to selector
The units for the selector are now based on what country is
available. Eventually they will be limited further for battle.
Added a factory function that can take the name and give me
a unit for easy army creation. Added first strike support in
the controller so that they go first and are excluded from
normal combat in round one after that.